backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.2
distro==1.9.0
faiss-cpu==1.10.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.2
ftfy==6.3.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
open_clip_torch==2.32.0
openai==1.73.0
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
wcwidth==0.2.13
